requirements.txt
fitz==0.0.1.dev2
fpdf==1.7.2
openai==1.50.0
pdf2image==1.17.0
Pillow==10.4.0
pydantic==2.9.2
pytesseract==0.3.13
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.38.0
